# TUI Performance Optimization