… packages/tui/internal/btree/btree_test.go
… packages/tui/internal/components/chat/batch_benchmark_test.go
… packages/tui/internal/components/chat/batch_processor.go
… packages/tui/internal/components/chat/editor.go
… packages/tui/internal/components/chat/editor_adaptive_test.go
… packages/tui/internal/components/chat/messages.go
… packages/tui/internal/components/chat/messages_benchmark_test.go
… packages/tui/internal/components/chat/sliding_window.go
… packages/tui/internal/components/chat/sliding_window_test.go
… packages/tui/internal/components/diff/diff.go
… packages/tui/internal/components/diff/diff_benchmark_test.go
… packages/tui/internal/components/diff/syntax_cache.go
… packages/tui/internal/components/textarea/textarea_adaptive.go
… packages/tui/internal/components/textarea/textarea_adaptive_benchmark_test.go
… packages/tui/internal/components/textarea/textarea_benchmark_test.go
… packages/tui/internal/components/textarea/textarea_rope.go
… packages/tui/internal/components/textarea/textarea_rope_benchmark_test.go
… packages/tui/internal/rangemap/rangemap.go
… packages/tui/internal/rangemap/rangemap_test.go
… packages/tui/internal/rope/benchmark_test.go
… packages/tui/internal/rope/rope.go
… packages/tui/internal/rope/rope_adapter.go
… packages/tui/internal/rope/rope_adapter_test.go
… packages/tui/internal/rope/rope_test.go